third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/simple-selector.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.expandElementsTree(onElementsTreeExpanded);

  function onElementsTreeExpanded() {
    var container = ElementsTestRunner.expandedNodeWithId('container');
    var children = container.children();
    for (var i = 0; i < children.length; ++i)
      TestRunner.addResult(Components.DOMPresentationUtils.simpleSelector(children[i]));
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>Verifies simple selector behavior.</p>
<div id="container" style="display: none">
    <div id="foo"></div>
    <div id="bar" class="foo"></div>
    <div class="title"></div>
    <input></input>
    <input type="text"></input>
    <input type="text" class="text-input"></input>
    <input type="text" id="text-input"></input>
    <p class="article"></p>
    <span></span>
    <b></b>
</div>

</body>
</html>
