third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/simple-selector-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies simple selector behavior.

div#foo
div#bar
.title
input
input[type="text"]
input.text-input
input#text-input
p.article
span
b
